backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djoser
drf-yasg
requests
pillow
